Fix team points and handicaps to use integer types
- Change Team.Points from float64 to int (league points are always integers)
- Change Handicaps from map[string]float64 to map[string]int (point deductions are always integers)
- Update all function signatures and type conversions accordingly
- Fix display formatting in demo.go to show points as integers
- Fix solver logging format specifier from %.1f to %d
- Group outright marks by market name in output for better readability

+++ pkg/outrights/api.go
@@ -59,7 +59,7 @@ func Simulate(events []Event, markets []Market, opts ...SimOptions) SimulationRe
 		Ratings:         make(map[string]float64),
 		TrainingSet:     trainingEvents,
 		Events:          predictionEvents,
-		Handicaps:       make(map[string]float64),
+		Handicaps:       make(map[string]int),
 		Markets:         markets,
 		PopulationSize:  8,
 		MutationFactor:  0.1,
@@ -278,7 +278,7 @@ func calcPPGRatings(teamNames []string, ratings map[string]float64, homeAdvantag
 	return ppgRatings
 }
 
-func calcExpectedSeasonPoints(teamNames []string, events []Event, handicaps map[string]float64, 
+func calcExpectedSeasonPoints(teamNames []string, events []Event, handicaps map[string]int, 
 	remainingFixtures []string, ratings map[string]float64, homeAdvantage float64) map[string]float64 {
 	
 	// Start with current league table points
@@ -286,7 +286,7 @@ func calcExpectedSeasonPoints(teamNames []string, events []Event, handicaps map[
 	expPoints := make(map[string]float64)
 	
 	for _, team := range leagueTable {
-		expPoints[team.Name] = team.Points
+		expPoints[team.Name] = float64(team.Points)
 	}
 	
 	// Add expected points from remaining fixtures
